Change Casbin functions to real async/await functions to make sure the user has the role before sending further requests. Add 2 middlewares in app.js to attach the app instance to req object in case of test environment Chnage the deregister delete request to use query instead of body

+++ src/tests/v1/registerRoute.test.js
@@ -5,7 +5,7 @@ const { generateMockUserRoute } = require('../../utils/generators');
 
 describe('POST /v1/register endpoint', () => {
 
-  let app, mockUser;
+  let app, mockUser, res;
 
   const headers = {
     'x-development-token': process.env.X_DEVELOPMENT_TOKEN,
@@ -13,7 +13,6 @@ describe('POST /v1/register endpoint', () => {
 
   // Utility function to create test requests
   const register = async (data) => await request(app).post('/v1/register').set(headers).send(data);
-  const deregister = async (data) => await request(app).delete('/v1/deregister').set(headers).send(data);
 
   beforeAll(async () => {
     app = await createApp();
@@ -21,9 +20,20 @@ describe('POST /v1/register endpoint', () => {
 
   // Clean up after each tests
   afterEach(async () => {
-    const res = await deregister(mockUser);
-    if (res.statusCode >= 400) {
-      await db.deleteUserByUsername(mockUser.username);
+    if (res.statusCode === 201) {
+      const userId = res.body.userId;
+      // Get the test user from the database
+      const createdUser = await db.getUserByUserId(userId);
+      const inactiveUser = { username: createdUser.username, activationCode: createdUser.activation_code };
+      await request(app).post('/v1/activate-by-code').set(headers).send(inactiveUser);
+      const user = { usernameOrEmail: mockUser.username, password: mockUser.password };
+      res = await request(app).post('/v1/login').set(headers).send(user);
+      const authData = res.body;
+      res = await request(app).del('/v1/deregister').set('Authorization', `Bearer ${authData.token}`);
+      if (res.statusCode >= 400) {
+        await db.deleteUserByUsername(mockUser.username);
+        console.log(`Couldn't delete the user: '${mockUser.username}'`);
+      }
     }
   });
 
@@ -34,7 +44,7 @@ describe('POST /v1/register endpoint', () => {
 
   it('should return 201 for successful registration', async () => {
     mockUser = generateMockUserRoute();
-    const res = await register(mockUser);
+    res = await register(mockUser);
     expect(res.statusCode).toBe(201);
     expect(res.body).toEqual({
       message: "User registered successfully.",
@@ -50,8 +60,9 @@ describe('POST /v1/register endpoint', () => {
 
   // Test Cases for Username
   it('should return 400 if username is not provided', async () => {
-    mockUser = { email: 'test@example.com', password: 'Password123!' };
-    const res = await register(mockUser);
+    mockUser = generateMockUserRoute();
+    delete mockUser.username;
+    res = await register(mockUser);
     expect(res.statusCode).toBe(400);
     expect(res.body.errors).toEqual(
       expect.arrayContaining([expect.objectContaining({ msg: 'Username is required.' })])
@@ -59,8 +70,9 @@ describe('POST /v1/register endpoint', () => {
   });
 
   it('should return 400 if username is too short', async () => {
-    mockUser = { username: 'test', email: 'test@example.com', password: 'Password123!' };
-    const res = await register(mockUser);
+    mockUser = generateMockUserRoute();
+    mockUser.username = 'test';
+    res = await register(mockUser);
     expect(res.statusCode).toBe(400);
     expect(res.body.errors).toEqual(
       expect.arrayContaining([expect.objectContaining({ msg: 'Username must be between 5 and 30 characters.' })])
@@ -68,8 +80,9 @@ describe('POST /v1/register endpoint', () => {
   });
 
   it('should return 400 if username contains invalid characters', async () => {
-    mockUser = { username: 'user*name', email: 'test@example.com', password: 'Password123!' };
-    const res = await register(mockUser);
+    mockUser = generateMockUserRoute();
+    mockUser.username = 'user*name';
+    res = await register(mockUser);
     expect(res.statusCode).toBe(400);
     expect(res.body.errors).toEqual(
       expect.arrayContaining([expect.objectContaining({ msg: 'Username can only contain letters, numbers, and underscores.' })])
@@ -80,9 +93,10 @@ describe('POST /v1/register endpoint', () => {
     const reservedUsernames = ['super', 'superdata', 'devhead', 'developer', 'saleshead', 'sales', 'support',
       'admin', 'admindata', 'officer', 'agent', 'enduser', 'public', 'administrator',
       'manager', 'staff', 'employee'];
+    mockUser = generateMockUserRoute();
     for (const username of reservedUsernames) {
-      mockUser = { username, email: 'test@example.com', password: 'Password123!' };
-      const res = await register(mockUser);
+      mockUser.username = username;
+      res = await register(mockUser);
       expect(res.statusCode).toBe(400);
       expect(res.body.errors).toEqual(
         expect.arrayContaining([expect.objectContaining({ msg: 'Username cannot be a reserved word.' })])
@@ -92,8 +106,9 @@ describe('POST /v1/register endpoint', () => {
 
   // Test Cases for Email
   it('should return 400 if email is not provided', async () => {
-    mockUser = { username: 'testuser', password: 'Password123!' };
-    const res = await register(mockUser);
+    mockUser = generateMockUserRoute();
+    delete mockUser.email;
+    res = await register(mockUser);
     expect(res.statusCode).toBe(400);
     expect(res.body.errors).toEqual(
       expect.arrayContaining([expect.objectContaining({ msg: 'Email is required.' })])
@@ -101,7 +116,9 @@ describe('POST /v1/register endpoint', () => {
   });
 
   it('should return 400 if email is invalid', async () => {
-    const res = await register(mockUser = { username: 'testuser', email: 'invalidemail', password: 'Password123!' });
+    mockUser = generateMockUserRoute();
+    mockUser.email = 'invalidemail';
+    res = await register(mockUser);
     expect(res.statusCode).toBe(400);
     expect(res.body.errors).toEqual(
       expect.arrayContaining([expect.objectContaining({ msg: 'Invalid email address.' })])
@@ -109,7 +126,9 @@ describe('POST /v1/register endpoint', () => {
   });
 
   it('should return 400 if email is not a string', async () => {
-    const res = await register(mockUser = { username: 'testuser', email: 4, password: 'Password123!' });
+    mockUser = generateMockUserRoute();
+    mockUser.email = 4;
+    res = await register(mockUser);
     expect(res.statusCode).toBe(400);
     expect(res.body.errors).toEqual(
       expect.arrayContaining([expect.objectContaining({ msg: 'Email must be a string.' })])
@@ -117,8 +136,9 @@ describe('POST /v1/register endpoint', () => {
   });
 
   it('should return 400 if email is passed as null', async () => {
-    mockUser = { username: 'testuser', email: null, password: 'Password123!' };
-    const res = await register(mockUser);
+    mockUser = generateMockUserRoute();
+    mockUser.email = null;
+    res = await register(mockUser);
     expect(res.statusCode).toBe(400);
     expect(res.body.errors).toEqual(
       expect.arrayContaining([expect.objectContaining({ msg: 'Email must be a string.' })])
@@ -127,8 +147,9 @@ describe('POST /v1/register endpoint', () => {
 
   it('should return 400 if email is too long', async () => {
     const longEmail = `${'a'.repeat(256)}@example.com`;
-    mockUser = { username: 'testuser', email: longEmail, password: 'Password123!' };
-    const res = await register(mockUser);
+    mockUser = generateMockUserRoute();
+    mockUser.email = longEmail;
+    res = await register(mockUser);
     expect(res.statusCode).toBe(400);
     expect(res.body.errors).toEqual(
       expect.arrayContaining([expect.objectContaining({ msg: 'Email must be between 5 and 255 characters.' })])
@@ -137,8 +158,9 @@ describe('POST /v1/register endpoint', () => {
 
   // Test Cases for Password
   it('should return 400 if password is not provided', async () => {
-    mockUser = { username: 'testuser', email: 'test@example.com' };
-    const res = await register(mockUser);
+    mockUser = generateMockUserRoute();
+    delete mockUser.password;
+    res = await register(mockUser);
     expect(res.statusCode).toBe(400);
     expect(res.body.errors).toEqual(
       expect.arrayContaining([expect.objectContaining({ msg: 'Password is required.' })])
@@ -146,8 +168,9 @@ describe('POST /v1/register endpoint', () => {
   });
 
   it('should return 400 if password is too short', async () => {
-    mockUser = { username: 'testuser', email: 'test@example.com', password: 'short' };
-    const res = await register(mockUser);
+    mockUser = generateMockUserRoute();
+    mockUser.password = 'short';
+    res = await register(mockUser);
     expect(res.statusCode).toBe(400);
     expect(res.body.errors).toEqual(
       expect.arrayContaining([expect.objectContaining({ msg: 'Password must be between 8 and 30 characters.' })])
@@ -155,8 +178,9 @@ describe('POST /v1/register endpoint', () => {
   });
 
   it('should return 400 if password does not contain uppercase letter', async () => {
-    mockUser = { username: 'testuser', email: 'test@example.com', password: 'password123!' };
-    const res = await register(mockUser);
+    mockUser = generateMockUserRoute();
+    mockUser.password = 'password123!';
+    res = await register(mockUser);
     expect(res.statusCode).toBe(400);
     expect(res.body.errors).toEqual(
       expect.arrayContaining([expect.objectContaining({ msg: 'Password must contain at least one uppercase letter.' })])
@@ -164,8 +188,9 @@ describe('POST /v1/register endpoint', () => {
   });
 
   it('should return 400 if password does not contain digit', async () => {
-    mockUser = { username: 'testuser', email: 'test@example.com', password: 'Password!' };
-    const res = await register(mockUser);
+    mockUser = generateMockUserRoute();
+    mockUser.password = 'Password!';
+    res = await register(mockUser);
     expect(res.statusCode).toBe(400);
     expect(res.body.errors).toEqual(
       expect.arrayContaining([expect.objectContaining({ msg: 'Password must contain at least one digit.' })])
@@ -173,8 +198,9 @@ describe('POST /v1/register endpoint', () => {
   });
 
   it('should return 400 if password does not contain special character', async () => {
-    mockUser = { username: 'testuser', email: 'test@example.com', password: 'Password123' };
-    const res = await register(mockUser);
+    mockUser = generateMockUserRoute();
+    mockUser.password = 'Password123';
+    res = await register(mockUser);
     expect(res.statusCode).toBe(400);
     expect(res.body.errors).toEqual(
       expect.arrayContaining([expect.objectContaining({ msg: 'Password must contain at least one latin symbol.' })])
@@ -183,8 +209,9 @@ describe('POST /v1/register endpoint', () => {
 
   // Test Cases for Display Name
   it('should return 400 if display name is too long', async () => {
-    mockUser = { username: 'testuser', email: 'test@example.com', password: 'Password123!', displayName: 'a'.repeat(51) }
-    const res = await register(mockUser);
+    mockUser = generateMockUserRoute();
+    mockUser.displayName = 'a'.repeat(51);
+    res = await register(mockUser);
     expect(res.statusCode).toBe(400);
     expect(res.body.errors).toEqual(
       expect.arrayContaining([expect.objectContaining({ msg: 'DisplayName can be a maximum of 50 characters.' })])
@@ -192,8 +219,9 @@ describe('POST /v1/register endpoint', () => {
   });
 
   it('should return 400 if display name is not a string', async () => {
-    mockUser = { username: 'testuser', email: 'test@example.com', password: 'Password123!', displayName: 4 }
-    const res = await register(mockUser);
+    mockUser = generateMockUserRoute();
+    mockUser.displayName = 4;
+    res = await register(mockUser);
     expect(res.statusCode).toBe(400);
     expect(res.body.errors).toEqual(
       expect.arrayContaining([expect.objectContaining({ msg: 'DisplayName must be a string.' })])
@@ -201,21 +229,24 @@ describe('POST /v1/register endpoint', () => {
   });
 
   it('should return 201 if display name is null', async () => {
-    mockUser = { username: 'testuser', email: 'test@example.com', password: 'Password123!', displayName: null }
-    const res = await register(mockUser);
+    mockUser = generateMockUserRoute();
+    mockUser.displayName = null;
+    res = await register(mockUser);
     expect(res.statusCode).toBe(201);
   });
 
   it('should return 201 if display name is an empty string', async () => {
-    mockUser = { username: 'testuser', email: 'test@example.com', password: 'Password123!', displayName: '' }
-    const res = await register(mockUser);
+    mockUser = generateMockUserRoute();
+    mockUser.displayName = '';
+    res = await register(mockUser);
     expect(res.statusCode).toBe(201);
   });
 
   // Test Cases for loginRedirectURL
   it('should return 400 if loginRedirectURL is invalid', async () => {
-    mockUser = { username: 'testuser', email: 'test@example.com', password: 'Password123!', loginRedirectURL: 'ftp://example.com' };
-    const res = await register(mockUser);
+    mockUser = generateMockUserRoute();
+    mockUser.loginRedirectURL = 'ftp://example.com';
+    res = await register(mockUser);
     expect(res.statusCode).toBe(400);
     expect(res.body.errors).toEqual(
       expect.arrayContaining([expect.objectContaining({ msg: 'The login redirect URL must be a valid URL starting with http:// or https://.' })])
@@ -223,8 +254,9 @@ describe('POST /v1/register endpoint', () => {
   });
 
   it('should return 400 if loginRedirectURL is not accessible', async () => {
-    mockUser = { username: 'testuser', email: 'test@example.com', password: 'Password123!', loginRedirectURL: 'https://thisurldoesnotexist1234.com' };
-    const res = await register(mockUser);
+    mockUser = generateMockUserRoute();
+    mockUser.loginRedirectURL = 'https://thisurldoesnotexist1234.com';
+    res = await register(mockUser);
     expect(res.statusCode).toBe(400);
     expect(res.body.errors).toEqual(
       expect.arrayContaining([expect.objectContaining({ msg: 'The login redirect URL is not a valid URL.' })])
